+++ standard/packages/core/src/component/ToolbarSection.tsx
@@ -15,6 +15,7 @@ export interface ToolbarContent {
 export interface ToolbarSectionProps extends ToolbarContent {
   name: string
   widgetGroups: ToolbarWidget[][]
+  titleId?: string
 }
 
 export class ToolbarSection extends BaseComponent<ToolbarSectionProps> {
@@ -52,7 +53,7 @@ export class ToolbarSection extends BaseComponent<ToolbarSectionProps> {
 
       if (buttonName === 'title') {
         children.push(
-          <h2 className="fc-toolbar-title">{props.title}</h2>,
+          <h2 className="fc-toolbar-title" id={props.titleId}>{props.title}</h2>,
         )
       } else {
         let isPressed = buttonName === props.activeButton