+++ standard/packages/core/src/component/ViewHarness.tsx
@@ -7,6 +7,8 @@ export interface ViewHarnessProps {
   heightLiquid?: boolean
   aspectRatio?: number
   children: ComponentChildren
+  label?: string
+  labelledBy?: string
 }
 
 export class ViewHarness extends Component<ViewHarnessProps> {
@@ -30,6 +32,8 @@ export class ViewHarness extends Component<ViewHarnessProps> {
             ? `${(1 / props.aspectRatio) * 100}%`
             : undefined
         }}
+        aria-label={props.label}
+        aria-labelledby={props.labelledBy}
       >
         {props.children}
       </div>